pyproject.toml
[project]
name = "problematic-source"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-magic>=0.4.27",
    "survey>=5.4.2",
    "zstandard>=0.25.0",
]

[build-system]

requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
